Conflicts:
	harem-web/src/main/webapp/WEB-INF/view/system/comm/groupTravel/detail.vm
	harem-web/src/main/webapp/WEB-INF/view/system/layout/empty.vm

+++ harem-biz/src/main/java/com/yimayhd/harem/service/HotelService.java
@@ -37,14 +37,14 @@ public interface HotelService {
     void modify(HotelVO hotelVO)throws Exception;
 
     /**
-     * 批量修改酒店状态
+     * 批量修改酒店资源状态
      * @param idList
      * @param hotelStatus
      * @throws Exception
      */
-    void setHotelStatusList(List<Integer> idList,int hotelStatus)throws Exception;
+    void setHotelStatusList(List<Long> idList,int hotelStatus)throws Exception;
     /**
-     * 修改酒店状态
+     * 修改酒店资源状态
      * @param id
      * @param hotelStatus
      * @throws Exception
